Refactored Answer for individual representation. Answers now each use HQ's official ID for reference.

+++ HackQ/Extensions.swift
@@ -148,7 +148,7 @@ extension Array where Element : StringProtocol
     }
 }
 
-extension CGFloat {
+extension Float {
     func percentageRoundedTo(places: UInt) -> String {
         return String(format: "%.\(places)f", self * 100)
     }
